+++ data/src/main/java/com/elta/android/data/features/auth/api/AuthApi.kt
@@ -13,6 +13,7 @@ import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.PUT
 
+@Suppress("TooManyFunctions", "ComplexInterface")
 interface AuthApi {
 
     @POST("api/auth/v1/accounts")