+++ data/src/main/java/com/elta/android/data/features/auth/datasource/AuthRemoteDataSource.kt
@@ -34,8 +34,8 @@ class AuthRemoteDataSource @Inject constructor(
         api.sendConfirmationLink().checkNetwork(checker)
 
     override fun sendResetPasswordLink(email: String): Completable =
-        api.sendPasswordResetLink(ResetPasswordLinkRequest(email))
+        api.sendPasswordResetLink(ResetPasswordLinkRequest(email)).checkNetwork(checker)
 
     override fun resetPassword(token: String, newPassword: String): Completable =
-        api.resetPassword(ResetPasswordRequest(token, newPassword))
+        api.resetPassword(ResetPasswordRequest(token, newPassword)).checkNetwork(checker)
 }