+++ domain/src/main/java/com/elta/android/domain/features/auth/model/SocialNetwork.kt
@@ -1,0 +1,5 @@
+package com.elta.android.domain.features.auth.model
+
+enum class SocialNetwork {
+    FB, VK, OK
+}