+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - [ELTA-111] - Added auth
 - [ELTA-112] - Added registration
 - [ELTA-114] - Added password recovery
+- [ELTA-113] - Added social network auth